move a utility function to utils.py
